issue(8): add rest endpoint to create leads
