+++ src/config/mongoose.js
@@ -1,0 +1,20 @@
+import mongoose from 'mongoose'
+import config from 'config/config'
+
+export function setup() {
+    mongoose.connection.on('connecting', () => {
+        console.log(`connecting to database: ${config.db}`)
+    })
+    mongoose.connection.on('connected', () => {
+        console.log(`connected to database: ${config.db}`)
+    })
+
+    mongoose.connect(config.db, { server: { socketOptions: { keepAlive: 1 } } })
+
+    mongoose.connection.on('error', () => {
+        throw new Error(`unable to connect to database: ${config.db}`)
+    })
+}
+
+export const Schema = mongoose.Schema
+export const model = mongoose.model.bind(mongoose)
