--- src/helpers/mongoose.js
@@ -1,3 +1,0 @@
-import mongoose from 'mongoose'
-export const Schema = mongoose.Schema
-export const model = mongoose.model.bind(mongoose)
