+++ src/routes/index.js
@@ -1,5 +1,6 @@
 import express from 'express'
-import leadsRoutes from '../components/leads/routes'
+import restify from './restify'
+// import leadsRoutes from 'components/leads/routes'
 
 const router = express.Router()    // eslint-disable-line new-cap
 
@@ -8,7 +9,8 @@ router.get('/health-check', (req, res) =>
     res.send('OK')
 )
 
+restify(router)
 // mount user routes at /users
-router.use('/leads', leadsRoutes)
+// router.use('/leads', leadsRoutes)
 
 export default router
