+++ src/routes/restify.js
@@ -1,0 +1,6 @@
+import restify from 'express-restify-mongoose'
+import { Lead } from 'components/models'
+
+export default function setup(router) {
+    restify.serve(router, Lead)
+}
